Function for decompressing an bmp from array
The function receives the next arguments: the display object, the x/y pos at which the picture will be drawn,the pointer to array,the size of the array,the width of the picture. The height is calculated automatically. For Pictures with more than 50% white background this function is even faster than the normal drawBitmap from Adafruit_SSD1306(+-10 ms).